Fix Issue #3.

+++ mcmm/analysis_visualization.py
@@ -19,8 +19,8 @@ class AnalysisViz(object):
     def _format_square(self, ax, minX, maxX, minY, maxY):
         ax.set_xlim(math.floor(minX), math.floor(maxX))
         ax.set_ylim(math.floor(minY), math.floor(maxY))
-        ax.set_xticks([i for i in range(math.floor(minX), math.ceil(maxX)+1)])
-        ax.set_yticks([i for i in range(math.floor(minY), math.ceil(maxY)+1)])
+        ax.set_xticks([i for i in range(int(math.floor(minX)), int(math.ceil(maxX)+1))])
+        ax.set_yticks([i for i in range(int(math.floor(minY)), int(math.ceil(maxY)+1))])
         ax.set_xlabel(r"$x$ / a.u.")
         ax.set_ylabel(r"$y$ / a.u.")
         ax.set_aspect('equal')
